+++ src/Common/CHANGELOG.md
@@ -7,6 +7,7 @@ The change log describes what is "Added", "Removed", "Changed" or "Fixed" betwee
 + Add to coordinates altitude
 + Add measuring distance between two coordinates
 + Fix potential issue in findPlaceByCoordinates method
++ Implement functionality for find neighbor places
 
 ## 2.0